🚚 engine.h: Extract `ActorPosition`
